chore: move libgit2 modules to core
