+++ lua/fugit2/view/components/commit_log_view.lua
@@ -6,7 +6,7 @@ local NuiText = require "nui.text"
 local Object = require "nui.object"
 local string_utils = require "plenary.strings"
 
-local git2 = require "fugit2.git2"
+local git2 = require "fugit2.core.git2"
 local pendulum = require "fugit2.core.pendulum"
 local utils = require "fugit2.utils"
