+++ lua/fugit2/view/git_status.lua
@@ -17,7 +17,7 @@ local LogView = require "fugit2.view.components.commit_log_view"
 local TreeBase = require "fugit2.view.components.base_tree_view"
 local UI = require "fugit2.view.components.menus"
 local fugit2_config = require "fugit2.config"
-local git2 = require "fugit2.git2"
+local git2 = require "fugit2.core.git2"
 local git_gpg = require "fugit2.core.git_gpg"
 local git_hooks = require "fugit2.core.git_hooks"
 local notifier = require "fugit2.notifier"
