+++ spec/fugit2/git2_spec.lua
@@ -1,5 +1,5 @@
-local git2 = require "fugit2.git2"
-local libgit2 = require "fugit2.libgit2"
+local git2 = require "fugit2.core.git2"
+local libgit2 = require "fugit2.core.libgit2"
 
 describe("git2", function()
   local repo --[[@as GitRepository]]
